chore: refactor collection list
